retrieving some game details when you click on a game
Creating getGameDetails function in http services to GET details about a game using its id. Creating multiple observables for info, trailers and screenshots and combining them into one observable.

Getting id of the game by subscribing to the activated route's parameters and getting the id parameter before passing it into getGameDetails()

Subscribing to these observables and setting the "game" variable to what was returned from the API. As a result, the info of a game is dynamic and variables are passed into the template

Creating ngOnDestroy() for details component so that all subscriptions are unsubscribed